Added new file for delete environment, addressed other comments.
