+++ src/libs/ajax/leonardo/Disks.ts
@@ -6,43 +6,39 @@ import { updatePdType } from 'src/pages/workspaces/workspace/analysis/utils/disk
 
 
 export const Disks = signal => {
-  const v2Func = () => {
-    const diskV2Root = 'api/v2/disks/'
-    return {
-      delete: (workspaceId, diskId): Promise<void> => {
-        return fetchLeo(`${diskV2Root}/${workspaceId}/${diskId}`, _.mergeAll([authOpts(), appIdentifier, { signal, method: 'DELETE' }]))
-      }
+  const diskV2Root = 'api/v2/disks/'
+  const v2Func = () => ({
+    delete: (workspaceId, diskId): Promise<void> => {
+      return fetchLeo(`${diskV2Root}/${workspaceId}/${diskId}`, _.mergeAll([authOpts(), appIdentifier, { signal, method: 'DELETE' }]))
     }
-  }
+  })
 
-  const v1Func = () => {
-    return {
-      list: async (labels = {}): Promise<ListDiskItem> => {
-        const res = await fetchLeo(`api/google/v1/disks${qs.stringify(labels, { addQueryPrefix: true })}`,
-          _.mergeAll([authOpts(), appIdentifier, { signal }]))
-        return res.json()
+  const v1Func = () => ({
+    list: async (labels = {}): Promise<ListDiskItem> => {
+      const res = await fetchLeo(`api/google/v1/disks${qs.stringify(labels, { addQueryPrefix: true })}`,
+        _.mergeAll([authOpts(), appIdentifier, { signal }]))
+      return res.json()
+    },
+    disk: (project: string, name: string) => ({
+      create: (props): Promise<void> => fetchLeo(`api/google/v1/disks/${project}/${name}`,
+        _.mergeAll([authOpts(), appIdentifier, { signal, method: 'POST' }, jsonBody(props)])
+      ),
+      delete: (): Promise<void> => {
+        return fetchLeo(`api/google/v1/disks/${project}/${name}`, _.mergeAll([authOpts(), appIdentifier, { signal, method: 'DELETE' }]))
       },
+      update: (size: number): Promise<void> => {
+        return fetchLeo(`api/google/v1/disks/${project}/${name}`,
+          _.mergeAll([authOpts(), jsonBody({ size }), appIdentifier, { signal, method: 'PATCH' }]))
+      },
+      details: async (): Promise<DecoratedPersistentDisk> => {
+        const res = await fetchLeo(`api/google/v1/disks/${project}/${name}`,
+          _.mergeAll([authOpts(), appIdentifier, { signal, method: 'GET' }]))
+        const disk: GetDiskItem = await res.json()
+        return updatePdType(disk)
+      }
+    })
+  })
 
-      disk: (project: string, name: string) => ({
-        create: (props): Promise<void> => fetchLeo(`api/google/v1/disks/${project}/${name}`,
-          _.mergeAll([authOpts(), appIdentifier, { signal, method: 'POST' }, jsonBody(props)])
-        ),
-        delete: (): Promise<void> => {
-          return fetchLeo(`api/google/v1/disks/${project}/${name}`, _.mergeAll([authOpts(), appIdentifier, { signal, method: 'DELETE' }]))
-        },
-        update: (size: number): Promise<void> => {
-          return fetchLeo(`api/google/v1/disks/${project}/${name}`,
-            _.mergeAll([authOpts(), jsonBody({ size }), appIdentifier, { signal, method: 'PATCH' }]))
-        },
-        details: async (): Promise<DecoratedPersistentDisk> => {
-          const res = await fetchLeo(`api/google/v1/disks/${project}/${name}`,
-            _.mergeAll([authOpts(), appIdentifier, { signal, method: 'GET' }]))
-          const disk: GetDiskItem = await res.json()
-          return updatePdType(disk)
-        }
-      })
-    }
-  }
   return {
     disksV1: v1Func,
     disksV2: v2Func,
